Now, the game works!

+++ Enemy.h
@@ -17,8 +17,8 @@ public:
 	Application *app;
 
 	SDL_Texture *graphics;
-	Point2d<float> position;
-	Point2d<float> speed;
+	Point2d<int> position;
+	Point2d<int> speed;
 	Animation anim;
 	Uint32 born;
 	Uint32 life;
